+++ fmt/src/inline_config.rs
@@ -1,4 +1,4 @@
-use crate::comments::{CommentState, CommentStringExt, Comments};
+use crate::comments::{CommentState, CommentStringExt};
 use itertools::Itertools;
 use solang_parser::pt::Loc;
 use std::{fmt, str::FromStr};
@@ -36,24 +36,6 @@ pub struct InvalidInlineConfigItem(String);
 impl fmt::Display for InvalidInlineConfigItem {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.write_fmt(format_args!("Invalid inline config item: {}", self.0))
-    }
-}
-
-impl Comments {
-    /// Parse all comments to return a list of inline config items. This will return an iterator of
-    /// results of parsing comments which start with `forgefmt:`
-    pub fn parse_inline_config_items(
-        &self,
-    ) -> impl Iterator<Item = Result<(Loc, InlineConfigItem), (Loc, InvalidInlineConfigItem)>> + '_
-    {
-        self.iter()
-            .filter_map(|comment| {
-                Some((comment, comment.contents().trim_start().strip_prefix("forgefmt:")?.trim()))
-            })
-            .map(|(comment, item)| {
-                let loc = comment.loc;
-                item.parse().map(|out| (loc, out)).map_err(|out| (loc, out))
-            })
     }
 }
